fix: change product ID type from Long to Integer in IProductService methods

+++ backend/src/main/java/com/farma_ya/service/IProductService.java
@@ -11,13 +11,13 @@ public interface IProductService {
 
     List<Product> getAllProducts();
 
-    Product getProductById(Long id);
+    Product getProductById(Integer id);
 
     Product createProduct(Product product);
 
-    Product updateProduct(Long id, Product productDetails);
+    Product updateProduct(Integer id, Product productDetails);
 
-    void deleteProduct(Long id);
+    void deleteProduct(Integer id);
 
     List<Product> searchProducts(String name);
